creacion modulo productos

+++ src/main/java/com/minsait/ejercicio_hexagonal/domain/productos/model/Producto.java
@@ -1,5 +1,7 @@
 package com.minsait.ejercicio_hexagonal.domain.productos.model;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
@@ -9,7 +11,7 @@ public class Producto {
 
     private final Long id;
     private final String nombre;
-    private final Double precio;
+    private final BigDecimal precio;
     private final Integer stock;
 
 }
